+++ src/main/java/br/com/Compasso/calculadora/repositorio/GrupoMedicamentoRepositorio.java
@@ -1,6 +1,7 @@
 package br.com.Compasso.calculadora.repositorio;
 
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -12,7 +13,7 @@ import br.com.Compasso.calculadora.modelo.GrupoMedicamentoEntity;
 
 
 @Repository
-public interface GrupoMedicamentoRepositorio extends JpaRepository<GrupoMedicamentoEntity, Long>{
+public interface GrupoMedicamentoRepositorio extends JpaRepository<GrupoMedicamentoEntity, BigInteger>{
 	
 	Optional<GrupoMedicamentoEntity> findByNome(String nome);
