+++ src/main/java/br/com/Compasso/calculadora/repositorio/LaboratorioRepositorio.java
@@ -1,6 +1,7 @@
 package br.com.Compasso.calculadora.repositorio;
 
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.Compasso.calculadora.modelo.LaboratorioEntity;
 
-public interface LaboratorioRepositorio extends JpaRepository<LaboratorioEntity, Long> {
+public interface LaboratorioRepositorio extends JpaRepository<LaboratorioEntity, BigInteger> {
 	
 		Optional<LaboratorioEntity> findByNome(String nome);
