+++ src/main/java/br/com/Compasso/calculadora/repositorio/MedicamentoRepositorio.java
@@ -1,6 +1,7 @@
 package br.com.Compasso.calculadora.repositorio;
 
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.Compasso.calculadora.modelo.MedicamentoEntity;
 
-public interface MedicamentoRepositorio extends JpaRepository<MedicamentoEntity, Long> {
+public interface MedicamentoRepositorio extends JpaRepository<MedicamentoEntity, BigInteger> {
 	Optional<MedicamentoEntity> findByNome(String nome);
